small fixes, forceNetworkFetch to true
