+++ app/core/core-ui/src/main/kotlin/com/hedvig/android/core/ui/appbar/TopAppbar.kt
@@ -1,13 +1,19 @@
 package com.hedvig.android.core.ui.appbar
 
 import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.WindowInsetsSides
+import androidx.compose.foundation.layout.displayCutout
+import androidx.compose.foundation.layout.only
+import androidx.compose.foundation.layout.union
 import androidx.compose.material.icons.Icons
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarColors
 import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -73,5 +79,55 @@ private inline fun TopAppBar(
       )
     },
     colors = TopAppBarDefaults.topAppBarColors(containerColor = containerColor),
+  )
+}
+
+@Composable
+fun TopAppBarWithBackAndClose(
+  title: String,
+  onNavigateUp: () -> Unit,
+  onClose: () -> Unit,
+  modifier: Modifier = Modifier,
+  windowInsets: WindowInsets = TopAppBarDefaults.windowInsets
+    .union(WindowInsets.displayCutout.only(WindowInsetsSides.Horizontal + WindowInsetsSides.Top)),
+  colors: TopAppBarColors = TopAppBarDefaults.topAppBarColors(
+    containerColor = MaterialTheme.colorScheme.background,
+    scrolledContainerColor = MaterialTheme.colorScheme.surface,
+  ),
+  scrollBehavior: TopAppBarScrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(),
+) {
+  TopAppBar(
+    modifier = modifier,
+    title = {
+      Text(
+        text = title,
+        style = MaterialTheme.typography.titleLarge,
+      )
+    },
+    navigationIcon = {
+      IconButton(
+        onClick = onNavigateUp,
+        content = {
+          Icon(
+            imageVector = Icons.Hedvig.ArrowBack,
+            contentDescription = null,
+          )
+        },
+      )
+    },
+    actions = {
+      IconButton(
+        onClick = onClose,
+        content = {
+          Icon(
+            imageVector = Icons.Hedvig.X,
+            contentDescription = null,
+          )
+        },
+      )
+    },
+    windowInsets = windowInsets,
+    colors = colors,
+    scrollBehavior = scrollBehavior,
   )
 }
