+++ app/feature/feature-insurances/src/main/kotlin/com/hedvig/android/feature/insurances/insurance/presentation/InsurancePresenter.kt
@@ -107,7 +107,7 @@ internal class InsurancePresenter(
       loadInsuranceData(
         getInsuranceContractsUseCase = getInsuranceContractsUseCaseProvider.provide(),
         getCrossSellsUseCase = getCrossSellsUseCaseProvider.provide(),
-        forceNetworkFetch = isRetryingIteration,
+        forceNetworkFetch = true,
       ).fold(
         ifLeft = {
           Snapshot.withMutableSnapshot {
